net/mlx4_core: Fix GEN_EQE accessing uninitialixed mutex
We occasionally see in procedure mlx4_GEN_EQE that the driver
tries to grab an uninitialized mutex.

This can occur in only one of 2 ways:
1. We are trying to generate an async event on an uninitialized slave.
2. We are trying to generate an async event on an illegal slave number
   ( < 0 or > persist->num_vfs) or an inactive slave.

To deal with possibility 1: move the mutex initialization from
specific slave initialization sequence in procedure mlx_master_do_cmd to
mlx4_multi_func_init() (so that the mutex is always initialized
for all slaves).

To deal with possibility 2: check in procedure mlx4_GEN_EQE that the
slave number provided is in the proper range and that the slave is active.

issue: 520893

Change-Id: I123af80e1fb1769c33053792af0a72b1c3573fb3
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>