net/mlx5_core: New init and exit flow for mlx5_core
In the new flow, we separate the pci initialization and teardown from the
initialization and teardown of the other resources.

init_one calls mlx5_pci_init that handles the pci resources initialization.
It then calls mlx5_load_one to initialize the remainder of the resources.

When removing a device, remove_one is invoked. However, now remove_one calls
mlx5_unload_one to free all the resources except the pci resources. When
mlx5_unload_one returns, mlx5_pci_close is called to free the pci resources.

The above separation will allow us to implement the pci error handlers and
suspend and resume callbacks.

issue: 440324
Change-Id: I20c5756cb3b73e454d0e772d18dd93af068de862
Signed-off-by: Majd Dibbiny <majd@mellanox.com>