cover letter [ib-next:admin_vf]
Under the existing implementation for virtual GIDs, if the SM is not reachable
or incurs a delayed response, or if the VF is probed into a VM before their
GUID is registered with the SM, there exists a window in time in which the VF
sees an incorrect GID, i.e., not the GID that was intended by the admin. This
results in exposing a temporal identity to the VF.

Moreover, a subsequent change in the alias GID causes a spec-incompliant change
to the VF identity. Some guest operating systems, such as Windows, cannot
tolerate such changes.

This series solves above problem by exposing the admin desired value instead of
the value that was approved by the SM. As long as the SM doesn't approve the
GID, the VF would see its link as down.

In addition, we request GIDs from the SM on demand, i.e., when a VF actually
needs them, and release them when the GIDs are no longer in use. In cloud
environments, this is useful for GID migrations, in which a GID is assigned to
a VF on the destination HCA, while the VF on the source HCA is shut down (but
the GID was not administratively released).

For reasons of compatibility, an explicit admin request to set/change a GUID
entry is done immediately, regardless of whether the VF is active or not. This
allows administrators to change the GUID without the need to unbind/bind the
VF.

In addition, the existing implementation doesn't support a persistency
mechanism to retry a GUID request when the SM has rejected it for any reason.
The PF driver shall keep trying to acquire the specified GUID indefinitely by
utilizing an exponential back off scheme, this should be managed per GUID and
be aligned with other incoming admin requests.

This ability needed especially for the on-demand GUID feature. In this case, we
must manage the GUID's status per entry and handle cases that some entries are
temporarily rejected.

The first patch adds the persistency support and is pre-requisites for the
series.  Further patches make the change to use the admin VF behavior as
described above.

Finally, the default mode is changed to be HOST assigned instead of SM
assigned. This is the expected operational mode, because it doesn't depend on
SM availability as described above.

Change-Id: I0e175eb2f3b4b1629e424260db82d3b18c95bd1f