net/mlx5_core: Add ISSI module parameter
This module parameter provide a WA to work with CX4 IB.

To work with CX4 IB, set it to 0.
To work with CX4 ETH, set it to 1.
On CIB, it doesn't matter.

This module parameter is temp and should be removed once the CX4 IB bug is
resolved.

issue: none
Change-Id: I1969aca328717cca3a920aac9cb4c5b4af7ec97c
Signed-off-by: Majd Dibbiny <majd@mellanox.com>