net/mlx5: Extend mlx5_core to support ConnectX-4 Ethernet functionality [net-next:mlx5e]
This is the Ethernet part of the driver for the Mellanox ConnectX(R)-4
Single/Dual-Port Adapter supporting 100Gb/s with VPI.  The driver
extends the existing mlx5 driver with Ethernet functionality.

This patch contains the driver entry points but does not include
transmit and receive (see the previous patch in the series) routines.
It also adds a new build option to Kconfig to enable/disable the
Ethernet functionality.

Currently, Kconfig is programmed to make Ethernet and Infiniband
functionality mutally exclusive. This limitation will be removed in the
very near future.

Issue: None
Change-Id: Id8e0a66733374a60f9e29ba83431337127601553
Signed-off-by: Amir Vadai <amirv@mellanox.com>