net/mlx4_core: Enhance the MAD_IFC wrapper to convert VF port to physical [net-next:single-port-ib-vf]
Single port VFs always provide port = 1 (even if the actual physical
port used is port 2). As such, we need to convert the port provided
by the VF to the physical port before calling into the firmware.

It turns out that the Linux mlx4 VF RoCE driver maintains a copy of
the GID table and hence this change became critical only for single
ported IB VFs, but it could be needed for other RoCE VF drivers too.

issue: 546607
Change-Id: Ia93ada8fe0360b2d12e87f9de4dcbc2dc7649901
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>