IB/mlx4: Replace mechanism for RoCE GID management [ib-next:roce_gid_table_mgmt]
Manage RoCE gid table with logic in IB/core, which is common to all
vendors, and remove the mechanism from the mlx4 IB driver.
Since management of the GID cache may lead to index mismatch with the
hardware GID table, a translation between indexes is required when
modifying a QP or creating an address handle.

Change-Id: I18ff6b4a65ef9ac954a0c2b906259b71f5ba9bf8
Signed-off-by: Moni Shoua <monis@mellanox.com>