IB/mlx4: Replace mechanism for RoCE GID management [ib-next:new_roce_gid_mgmt]
Manage RoCE gid table with logic in IB/core, which is common to all
vendors, and remove the mechanism from the mlx4 IB driver.
Since management of the GID cache may lead to index mismatch with the
hardware GID table, a translation between indexes is required when
modifying a QP or creating an address handle.

Change-Id: If9a60278483fb19be3b29fc20eb5035d05a12ac3
Signed-off-by: Moni Shoua <monis@mellanox.com>