IB/mlx4: Convert slave port before building a derived address-handle [net-next:single-port-ib-vf]
When multiplexling a MAD sent from guest, we should convert the port used
by the guest to send the packet to the actual physical port which will be
use to transmit the packet before building the relevant address-handle (AH).

This is needed under VPI for single ported VFs since the code that builds
the AH (mlx4_ib_query_ah()) makes decisions based on the input port. If we
use the port number provided by the guest, it could have different protocol
vs. the one this packat has to go from and hence the result could be wrong.

So far the conversion was done after the AH was built and it worked for
single ported Eth VFs which are not enabled under VPI. When adding support
for single ported IB VFs and VPI, we hit that.

issue: 546607
Change-Id: Ie9091f5a2cc407532dfd18f3f2ec80c65aa87876
Fixes: 449fc48866f7 ('net/mlx4: Adapt code for N-Port VF')
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>