IB/core: Add gid_type to gid attirbute
In order to support multiple GID types, we need to store
the gid_type with each GID.

This implies that gid_type should be added to roce_gid_table
metadate, as well as adding the required support in rdma_cm.

When using rdma cm, we want to take the gid_type from
the rdma_id_private. This is mandatory before adding
an API from user-space/configfs that sets
the gid_type of CM connection.

Change-Id: I022cdcf7eaaf4977654132b8f14c1ca8e3f659d1
Signed-off-by: Matan Barak <matanb@mellanox.com>