net/mlx5e: Allocate DMA coherent memory on reader NUMA node [net-next:ps1]
By affinity hints and XPS, each mlx5e channel is assigned a CPU
core.

Channel DMA coherent memory that is written by the NIC and read
by SW (e.g CQ buffer) is allocated on the NUMA node of the CPU
core assigned for the channel.

Channel DMA coherent memory that is written by SW and read by the
NIC (e.g SQ/RQ buffer) is allocated on the NUMA node of the NIC.

Doorbell record (written by SW and read by the NIC) is an
exception since it is accessed by SW more frequently.

Change-Id: I1bbe5a9f94d81ebecfa5e4bf8ca287cbb129f4e0
Signed-off-by: Saeed Mahameed <saeedm@mellanox.com>