IB/mlx4: Replace mechanism for RoCE GID management [ib-next:roce_gid_mgmt_v6]
Manage RoCE gid table with logic in IB/core, which is common to all
vendors, and remove the mechanism from the mlx4 IB driver.
Since management of the GID cache may lead to index mismatch with the
hardware GID table, a translation between indexes is required when
modifying a QP or creating an address handle.

Change-Id: I439d25bee4d6265fcb4566b60db444a04d7b819b
Signed-off-by: Moni Shoua <monis@mellanox.com>