net/mlx4_core: Fix wrong index in propagating port change event to VFs [net-next:bug-fix]
The port-change event processing in procedure mlx4_eq_int() uses "slave"
as the vf_oper array index. Since the value of "slave" is the PF function
index, the result is that the PF link state is used for deciding to
propagate the event for all the VFs. The VF link state should be used,
so the VF function index should be used here.

issue: 561232

Change-Id: I20e96170c5c96f36621d3f18a8cd738bf6504d25
Fixes: 948e306d ('net/mlx4: Add VF link state support')
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>
Signed-off-by: Matan Barak <matanb@mellanox.com>