net/mlx5_core: Fix the set ISSI flow to handle wrong firmware behavior [net-next:fix_issi]
The firmware doesn't publish whether it supports the query_issi command or not,
so in order to know, we invoke it unconditionally, and if it fails with status
BAD_OPCODE, we are supposed to continue working with ISSI 0.

Some older firmware versions behave wrong and don't return BAD_OPCODE in the status. To
cope with these firmwares, we don't check the status when the command fails, and
continue working with ISSI 0.

issue: none
Change-Id: I90025cb5c8b9d05bd9257b4d6cd7e6706cd18a85
Signed-off-by: Majd Dibbiny <majd@mellanox.com>