net/mlx5_core: Add RQ and SQ event handling
We now use RQ/SQ to implement IB verbs QPs, So the IB QP affiliated
events are ow affiliated also with SQs and RQs.

Since SQ, RQ and QP resource number do not share the same name
space, a queue type field was added to the event data.  We use it
here to locate the relevant resource SW object to handle the event.

Change-Id: Iebe624fb2e4e9cd9469aa437f18e18736667626c
Signed-off-by: Majd Dibbiny <majd@mellanox.com>