net/mlx5_core: Add RQ and SQ event handling
RQ/SQ will be used to implement IB verbs QPs, so the IB QP affiliated
events are affiliated also with SQs and RQs.

Since SQ, RQ and QP resource numbers do not share the same name
space, a queue type field was added to the event data to specify
the SW object that the event is affiliated with.

issue: none
Change-Id: I9d6928f64ae2846958111aa73d67991d49d0bde7
Signed-off-by: Majd Dibbiny <majd@mellanox.com>