net/mlx5e: Replace TIR inline RQ with pointer to an RQT with single RQ
In the following commits we will change the ndo_stop flow such that it
will not close all mlx5e priv resources but rather only re-direct the
RX traffic to a "drop_rq" (an RQ that silently drops, at the NIC HW
level, all traffic).

In order for this re-direction to involve only MODIFY_RQT operation
and not to involve an additional MODIFY_TIR operation, we replace
the TIR inline RQ with a pointer to an RQT with single RQ, which is
logically equivalent.

Change-Id: I333c55fdae25e4c05d06c07b878d6fa5b13c692b
Signed-off-by: Achiad Shochat <achiad@mellanox.com>

Conflicts:
	drivers/net/ethernet/mellanox/mlx5/core/en_main.c