net/mlx5e: Rename/move functions following the ndo_stop flow change
Rename some functions that used to be invoked upon ndo_open/stop and
are now invoked upon create/destroy_netdev() in order to better hint
their place in the flow.

Change some functions location in the file so that functions involved
in ndo_open/stop flow will not be interleaved with other functions.

This is a cosmetic change, no logical change here.

Change-Id: Ia7f3479c274ac8de8093a3a48a35c032ce46860b
Signed-off-by: Achiad Shochat <achiad@mellanox.com>