IB/core: Remove smac and vlan id from path record
The GID cache accompanies every GID with attributes.
The GID attributes link the GID with its netdevice, which could be
resolved to smac and vlan id easily. Since we've added the netdevice
(ifindex and net) to the path record, storing the L2 attributes is
duplicated data and hence these attributes are removed.

Change-Id: Ic935937fd69531a0fa1bd026e8ec6b3ecf78e46c
Signed-off-by: Matan Barak <matanb@mellanox.com>