net/mlx5e: Add switchdev fdb (l2) offload
Add switchdev object offloads entries to the VF representor netdevice
and support the FDB (l2) object type.

Currently this is limited to one mac only, TBD extending
it to multiple macs.

Also, currently the linux bridge will not attepmt to offload FDB
forwarding into the port device and hence to actually use it one
needs to run:

	$bridge fdb add MAC dev DEV self

on the representor device with the VF mac

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>