net/mlx5_core: Introduce flow steering API
The following relations are supported (every relation is one-to-many):
namespace --> priorities --> flow-tables --> flow-groups -->
flow-entries --> destinations
--> namespace

Namespace are used in order to isolate different usages of steering
(for example, downstream patches will add a namespace for the NIC
driver and for mlx5_ib usages).
Priorities are a way in order to maintain matching order. For
example, entries in priority i are matched before entries in
priority i+1.
Each priority has either flow-tables or namespaces (but not both).

Flow table are a way to group several steering entries. The order
in each flow-table (when auto-grouping isn't used) is undefined.
Each flow-table contains a few flow groups, which defines the
matching fields.

Change-Id: I4cb3c050803d65fffd973583772705eeaa85b982
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>