net/mlx5_core: Flow steering initialization
Flow steering initialization is based on static tree which
illustrates the flow steering tree when the driver is loaded. The
initilization consider the max supported flow table level of the device,
minimum of 2 kernel flow tables(vlan and mac) is required to have
kernel flow table functionality.
In addition this patch calculate prio->max_ft recursively.

Change-Id: I843aec237c60883c2c654ef872ebb7279cb2da73
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>