IB/core: Add dont trap flow create flag
Don't trap flag indicates that QP will recive traffic, but will not
steal it.
When a packet matches a flow steering rule that was created with
DONT_TRAP attribute, the QPs assigned to this rule will get this
packet, but matching will continue to other similar/lower priority
rules. Meaning, if these rules have other QPs assigned to them,
they'll get this packet too.
If both DONT_TRAP and regular trap rules exist in the same priority
matching the same packet, the behavior is undefined.

Issue: 552474
Change-Id: I8b69df9a9c4a1af95d0a6452f65d407e3fbbe863
Signed-off-by: Marina Varshaver <marinav@mellanox.com>