switchdev: Avoid long stack dumps on commit phase failure
Currently, mlx5 does all the required checks to determine if a given
flow can be actually offloaded in the commit and not prepare phase.

This results in errors we return on the commit phase and switchdev
screams, as a WA hack to avoid that, do warn only if the returned
driver error code is != -EOPNOTSUPP .

This will be changed when we move to use >= 4.3 kernels which
have the means to pass a driver cookie from the prepare to
the commit phase.

Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>