IB/cma: Print warning on different inner and header P_Keys
Patch 4c21b5bcef73 ("IB/cma: Add net_dev and private data checks to RDMA
CM") added checks for incoming RDMA CM requests that they can be matched to
a netdev based on the P_Key in the BTH of the request. This behavior was
reverted in ab3964ad2acf ("IB/cma: Use inner P_Key to determine netdev"),
since the mlx5 and ipath drivers didn't send the correct value in the BTH
P_Key.

Since the mlx5 and ipath drivers can now send GSI packets on different
P_Keys, we could revert the patch to let the rdma_cm module look on the BTH
P_Key when deciding to what netdev a packet belongs. However, that still
breaks compatibility with the older drivers.

Change the behavior to print a warning when receiving a request that has a
different BTH P_Key and inner payload P_Key. In the future, after users
have seen the warnings and upgraded their setups, remove the warning and
block these requests.

Issue: 578240
Change-Id: Iabafe33fead834a092f235150ee6582e20c2f266
Signed-off-by: Haggai Eran <haggaie@mellanox.com>