net/mlx4_en: Explicitly set no vlan tags in WQE ctrl segment when no vlan is present
commit f108c5578ca705 ("net/mlx4_en: Add support for hardware accelerated 802.1ad vlan")
neglected to set the ins_vlan field to zero when no vlan id is present
in the packet.
Since WQEs in the TX ring are not zeroed out between uses, this oversight
could result in having vlan flags present in the WQE ctrl segment when
no vlan is preset.

Fix was suggested by Gideon Naim of Mellanox.

issue: 634231

Change-Id: Ieba7c5c3460614eb4e10a8193702b44dc29003a2
Signed-off-by: Jack Morgenstein <jackm@dev.mellanox.co.il>