IB/mlx5: Add per port counters
In order to support statistics for ports, we attach
each QP to a counter set which is dedicate to this port.

Several QP types don't have PORT attribute in RESET to INIT state.
For this types, we assume the port was given in the creation.
This is mandatory in order to set the QP counters correctly.

Change-Id: If5fea47c7c886b62efd88db9aa6defbcfd081e21
Signed-off-by: Matan Barak <matanb@mellanox.com>