net/mlx5_core: Introduce flow steering API
The following relations are supported (every relation is one-to-many):
namespaces --> priorities -->flow-tables --> flow-groups -->
flow-entries --> destinations

Namespace are used in order to isolate different usages of
steering (for example, downstream patches will add a namespace for the
NIC driver and for mlx5_ib usages).
Priorities are a way in order to maintain matching order.
For example, entries in priority i are matched before entries
in priority i+1.
Each priority has either flow-tables or namespaces (but not both).
Flow tables are a way to group several steering entries.
Each flow-table contains a few flow groups, which defines
the matching fields.

Change-Id: I14fd1d335687572ea2469aacdf9ea4a91b658204
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>