net/mlx5: Configure cache line size for start and end padding
There is a hardware feature that will pad the start or end of a DMA to
be cache line aligned to avoid RMWs on the last cache line. The default
cache line size setting for this feature is 64B. This change configures
the hardware to use 128B alignment on systems with 128B cache lines.

Issue: 596607
Change-Id: I2acd4a4eb539aed810880f86801bcacad6dbc389
Signed-off-by: Daniel Jurgens <danielj@mellanox.com>