ovs: Add support for VXLAN source vport
When the source vport type is vxlan, use fib_lookup to find the
netdev that should receive such packet.
The fact the source port was VXLAN is stored in
flow->flow.key.misc.tunnel_type.
Get the udp dst_port from the vxlan vport, store it in tun->tp_dst
and set the relvent mask

Change-Id: Ic873ccb438667cc5311c652f4692148a8e250972
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>