net/mlx5e: Support offloading vlan push/pop actions
The CX4/4Lx HW can't push/pop vlans per flow.

We use the SRIOV VST mode as work-around. In VST, the SRIOV eSwitch
can tag packets sent from the VF, and strip packets sent to the VF.

Note that for "incoming" flows, OVS matches also on the vlan as part
of the packet 12-tuple fields. This works OK on the HW, since the eSwitch
steering is done before stripping the tag.

On slow path, when a packet arrives from a VF with vlan tag and that
vport is in VST, we remove the vlan before calling into the PF kernel.

Issue: 681037
Change-Id: I8218f53b9fa11801758eb134ebfa9af1b5d0630d
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>