net/mlx5_core: Managing root flow table
The root Flow Table for each Flow Table Type is defined,
by default, as the Flow Table with level 0.
When we create new flow table with level lower than the current
lowest flow table or it is the first flow table,
we update the FW about the new root flow table.

In order not to use empty flow tables and introduce new hops,
but still preserve space for flow-tables that have a priority
greater than the current flow table, we introduce this new set
root flow table command.
This command tells the HW to start matching the packets from the
assigned root flow table.

Issue: 665160
Change-Id: Ida9a208798c57f5803c8e2cdafe52bb3bef710fb
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Matan Barak <matanb@mellanox.com>