net/mlx5_core: Set start level and max_ft to priority
Traversing the tree nodes by DFS and set start level and max_ft
to each priority.
Start level depends on the max flow tables of the prior priorities
in the tree.

The leaves of the trees have max_ft set in them. Each node accumulates
the max_ft of its children and set it accordingly.

Issue: 665160
Change-Id: Ib3457772c9db2c2b65bdce8f96bd5e42627d6bcf
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Matan Barak <matanb@mellanox.com>