net/mlx5_core: Connect flow tables
Flow tables from different priorities should be chained together.
When packet is arrived we search for a match in the by-pass flow
tables(first we search for a match in priority 0 and if we don't find
we move to the next priority), if there is no match in all the by-pass
flow tables we move to search for a match in the kernel flow tables.

set the miss flow table in a new flow table to be the next one in
the list (via create flow table API). If we want to change an existing
flow table, for example in order to point from an existing flow table
to the new next-in-list flow table, we use the modify flow table API.

Issue: 665160
Change-Id: I0dc4c17811978451e60c835cc2d4211b1196d17a
Signed-off-by: Maor Gottlieb <maorg@mellanox.com>
Signed-off-by: Moni Shoua <monis@mellanox.com>
Signed-off-by: Matan Barak <matanb@mellanox.com>