net/mlx5e: Avoid storms when multicasting over representors
Currenlty multicast packets sent by the PF to representors
are loopbacked to the PF TIRs and this causes multicast storms.

Multicast loopback prevention is based on the SQ -> TIS -> TDN
chain to have the same TDN (Transport Domain ##) as the
RQ -> TIR -> TDN for the RQs of the receiving entity.

The SQs used by the representors do have this property, however,
after creating them we didn't follow the directives in commit
84f173f ('net/mlx5e: Added self loopback prevention') and the
PF TIRs were not updated and hence the loopback and storms
of the form pf->rep->pf->rep->... , fix that.

Issue:641107
Change-Id: I6b7a377a3b3cea8cf36fbde35e94e16754669d85
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>