net/mlx5e: Move rep init/cleanup code into ndo_{open/close}
Previously the representor netdevs were registered before
their SQs were created which resulted in a crash if a packet
was sent too early.
The rep channels are now created inside ndo_open similar to
the way it is done in other netdevs.
The insertion of steering rules that are relevant to the representors
was also moved to ndo_open.

Issue: 650339
Change-Id: Ib4934aa2af0dc558927d7295afe73890d45d2c5e
Signed-off-by: Ilya Lesokhin <ilyal@mellanox.com>